add actor factory
