+++ calvin/actor/actor.py
@@ -551,7 +551,7 @@ class Actor(object):
         return state
 
     @verify_status([STATUS.LOADED, STATUS.READY, STATUS.PENDING])
-    def _set_state(self, state):
+    def set_state(self, state):
         # Managed state handling
 
         # Update since if previously a shadow actor the init has been called first
@@ -601,7 +601,7 @@ class Actor(object):
         return self.state()
 
     def deserialize(self, data):
-        self._set_state(data)
+        self.set_state(data)
 
     def exception_handler(self, action, args, context):
         """Defult handler when encountering ExceptionTokens"""
