+++ calvin/runtime/north/metering.py
@@ -97,7 +97,7 @@ class Metering(object):
             _log.debug("get_timed_meter: User id not found")
             raise Exception("User id not found")
         t = time.time()
-        response = {actor_id: [d for d in data if d[0] > self.users[user_id]] 
+        response = {actor_id: [d for d in data if d[0] > self.users[user_id]]
                                 for actor_id, data in self.actors_log.iteritems()}
         self.users[user_id] = t
         self.forget(t)
